+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 * Add "naive" and "top-k" methods in MapieClassifier
 * Include J+aB method in regression tutorial
 * Add MNIST example for classification
+* Add cross-conformal for classification
 
 0.3.1 (2021-11-19)
 ------------------